Fix t3409.5
The idea was to test whether the rebase of M onto A2 (while preserving
the merge commits) in this graph works:

	A1 - B1 - M
	   \    /
	     A2

The original problem fixed in 12bf8283(rebase -i -p: include
non-first-parent commits in todo list) was that B1 was dropped.

However, the test was overzealous in assuming that M needs to be
preserved: when rebasing, in effect, B1 and M onto A2, B1 should
actually be cherry-picked, and M no longer be necessary. This is the
behavior of the upcoming new 'rebase -i -p'.

So let's test that the relevant commits are all there, but skip the test
whether the merge is still there (that would need to be a forced
non-fast-forward merge if B1 is rebased).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>