Fix t3414.4
This test case assumed that preserving merges would preserve a certain
set of merged branches, but it should have allowed for more commits to
be rebased.

In particular, the scenario was this:

	A - B - E - F - G
	  \   \       /
	    H   C - D

and after rebasing G from C onto H, it expected the history to look like
this:

	A - B - E - F - G
	  \           /
            H - D' --

while the upcoming 'rebase -i -p' will write it into:

	A - H - E' - F' - G
	      \         /
	        D' ----

Since B is reachable from C, it should not have been part of the rebased
history, therefore the original test whether the first parent of G is the
original F is arguably wrong.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>